Fixes and protections in user controller/endpoints

+++ backend/src/user/user.controller.ts
@@ -77,12 +77,14 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('grant-admin')
 	public async grantAdmin(@Query('login') login: string, @UserId() id: number): Promise<User[]>{
+		if (!login) return;
 		const executor: User = await this.service.getUser(id)
 		if (!executor) return;
 		if (executor.login == login) return [] as User[];
 		const hasExecutorPrivileges: boolean = executor.userRole >= 5 ? true : false
 		if (!hasExecutorPrivileges) return [] as User[]
 		const target: User = await this.service.getUserByLogin(login)
+		if (!target) return ;
 		const isTargetOwner: boolean = target.userRole >= 6 ? true : false
 		if (isTargetOwner) return [] as User[]
 		const isTargetBanned: boolean = target.isBanned ? true : false
@@ -93,12 +95,15 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('remove-admin')
 	public async removeAdmin(@UserId() id: number, @Query('login') login: string): Promise<User[]>{
+		if (!login) return;
 		const executor: User = await this.service.getUser(id)
 		if (!executor) return;
 		if (executor.login == login) return [] as User[];
 		const hasExecutorPrivileges: boolean = executor.userRole >= 5 ? true : false
 		if (!hasExecutorPrivileges) return [] as User[]
-		const isTargetOwner: boolean = (await this.service.getUserByLogin(login)).userRole >= 6 ? true : false
+		const target: User = await this.service.getUserByLogin(login);
+		if (!target) return;
+		const isTargetOwner: boolean = target.userRole >= 6 ? true : false
 		if (isTargetOwner) return [] as User[]
 		return this.service.removeAdmin(login);
 	}
@@ -106,12 +111,15 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('ban')
 	public async banUserFromWebsite(@UserId() id: number, @Query('login') login: string): Promise<User[]>{
+		if (!login) return
 		const executor: User = await this.service.getUser(id)
 		if (!executor) return;
 		if (executor.login == login) return [] as User[];
 		const hasExecutorPrivileges: boolean = executor.userRole >= 5 ? true : false
 		if (!hasExecutorPrivileges) return [] as User[]
-		const isTargetOwner: boolean = (await this.service.getUserByLogin(login)).userRole >= 6 ? true : false
+		const target: User = await this.service.getUserByLogin(login)
+		if (!target) return
+		const isTargetOwner: boolean = target.userRole >= 6 ? true : false
 		if (isTargetOwner) return [] as User[]
 		return this.service.banUserFromWebsite(login);
 	}
@@ -119,12 +127,15 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('unban')
 	public async removeBanUserFromWebsite(@UserId() id: number, @Query('login') login: string): Promise<User[]>{
+		if (!login) return
 		const executor: User = await this.service.getUser(id)
 		if (!executor) return;
 		if (executor.login == login) return [] as User[];
 		const hasExecutorPrivileges: boolean = executor.userRole >= 5 ? true : false
 		if (!hasExecutorPrivileges) return [] as User[]
-		const isTargetOwner: boolean = (await this.service.getUserByLogin(login)).userRole >= 6 ? true : false
+		const target: User = await this.service.getUserByLogin(login)
+		if (!target) return
+		const isTargetOwner: boolean = target.userRole >= 6 ? true : false
 		if (isTargetOwner) return [] as User[]
 		return this.service.removeBanUserFromWebsite(login);
 	}  
@@ -133,6 +144,7 @@ export class UserController {
 	@Get('my-blocked')
 	public async getMyBlockedUsers(@UserId() id: number): Promise<Array<string>>{
 		const user: User = await this.getUser(id)
+		if (!user) return [];
   		const bannedUsers: User[] = await this.service
   	  		.getBannedUsersByLogin(user.login)
 		return bannedUsers.map(m => m.login)
@@ -141,7 +153,10 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('block')
 	public async blockUser(@UserId() id: number, @Query('login') login: string): Promise<Array<string>>{
+		if (!login) return;
 		const user: User = await this.getUser(id)
+		if (!user) return
+		if (user.login == login) return
 		const blockUser: boolean = await this.chatService.banUser2User(user.login, login)
 		return await this.getMyBlockedUsers(id)
 	}
@@ -149,7 +164,10 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('unblock')
 	public async unBlockUser(@UserId() id: number, @Query('login') login: string): Promise<Array<string>>{
+		if (!login) return;
 		const user: User = await this.getUser(id)
+		if (!user) return
+		if (user.login == login) return
 		const unBlockUser: boolean = await this.chatService.noBanUser2User(user.login, login)
 		return await this.getMyBlockedUsers(id)
 	}
@@ -157,8 +175,10 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Get('friendshiprequest')
 	public async requestFriendship(@UserId() id: number, @Query('login') login: string): Promise<any>{
-		console.log("yee")
+		if (!login) return 
 		const user: User = await this.getUser(id)
+		if (!user) return
+		if (user.login == login) return
 		const friendshipSent: boolean = await this.service.requestFriendship(user.login, login)
 		return friendshipSent
 	}
@@ -166,42 +186,53 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Post('/friendship/accept')
 	public async acceptFriendship(@UserId() id: number, @Query('login') login: string): Promise<Array<string>>{
+		if (!login) return 
 		const user: User = await this.getUser(id)
+		if (!user) return
+		if (user.login == login) return
 		return await this.service.acceptFriendship(user.login, login)
 	}
 
 	@UseGuards(AuthGuard)
 	@Post('friendship/request/reject')
 	public async rejectFriendshipRequest(@UserId() id: number, @Query('login') login: string): Promise<Array<string>>{
+		if (!login) return 
 		const user: User = await this.getUser(id)
+		if (!user) return
+		if (user.login == login) return
 		return await this.service.rejectFriendshipRequest(user.login, login)
 	}
 
 	@UseGuards(AuthGuard)
 	@Post('friendship-remove')
 	public async removeFriendship(@UserId() id: number, @Query('login') login: string): Promise<Array<string>>{
-		console.log("yee")
 		const user: User = await this.getUser(id)
+		if (!user) return
+		if (user.login == login) return
 		return await this.service.removeFriendship(user.login, login)
 	}
 
  
-
+	@UseGuards(AuthGuard)
 	@Get('all')
-	public findAll(): Promise<User[]> {
-		return this.service.getAllUsers();
+	public async findAll(): Promise<User[]> {
+		return await this.service.getAllUsers();
 	}
 
 	@UseGuards(AuthGuard)
 	@Get(':id')
-	public getUser(@Param('id', ParseIntPipe) id: number): Promise<User>{
-		return this.service.getUser(id);
+	public async getUser(@Param('id', ParseIntPipe) id: number): Promise<User>{
+		if (!id) return;
+		return await this.service.getUser(id);
 	}
 
 	@UseGuards(AuthGuard)
 	@Get('first-login/:login')
-	public async isMyFirst(@Param('login') login: string): Promise<boolean> {
+	public async isMyFirst(@UserId() id: number, @Param('login') login: string): Promise<boolean> {
+		if (!login) return false
 		const user: User = await this.service.getUserByLogin(login);
+		if (!user) return false;
+		if (user.id != id) return false
 		if (user.firstLogin){
 			user.firstLogin = false;
 			await this.service.saveUser(user)
@@ -213,49 +244,53 @@ export class UserController {
 	@UseGuards(AuthGuard)
 	@Get('id/:login')
 	public getUserIdByLogin(@Param('login') login: string): Promise<number> {
+		if (!login) return
 		return this.service.getUserIdByLogin(login);
 	}
 
 	@UseGuards(AuthGuard)
 	@Get(':login/games')
-	public getGamesByUser(@Param('login') login: string): Promise<any> {
-		return this.service.getGamesByUser(login)
+	public async getGamesByUser(@Param('login') login: string): Promise<any> {
+		if (!login) return true;
+		return await this.service.getGamesByUser(login)
 	}
 
 	@UseGuards(AuthGuard)
 	@Get(':id/achievements')
-	public getUserAchievements(@Param('id', ParseIntPipe) id: number): Promise<Achievement[]> {
-	  return this.service.getUserAchievements(id);
+	public async getUserAchievements(@Param('id', ParseIntPipe) id: number): Promise<Achievement[]> {
+		if (!id) return [] as Achievement[];
+	  return await this.service.getUserAchievements(id);
 	}
 
 	//we should remove this too
-	@Post()
-	public createUser(@Body() body: CreateUserDto): Promise<User>{
-		return this.service.createUser(body);
+//	@Post()
+//	public createUser(@Body() body: CreateUserDto): Promise<User>{
+//		return this.service.createUser(body);
 
-	}
+//	}
 
 	/* remove this method! only for testing purposes 
 	*	is a way to test different users while we don't have a custom
 	*	signup and login method
 	* */
-	@Get('/force/:login')
-	public async getTokenFromLogin(@Param('login') login: string): Promise<any>{
-		const user: User = await this.service.getUserByLogin(login);
-		const payloadToSign = {login: login, id: user.id, role: user.userRole}
-		const access_token = await this.jwtService.signAsync(payloadToSign);
-		const decoded: JwtPayload = this.jwtService.decode(access_token) as JwtPayload;
-		return {
-			access_token: access_token,
-			expires_at: decoded.exp * 1000, //ms
-		};
-	}
+//	@Get('/force/:login')
+//	public async getTokenFromLogin(@Param('login') login: string): Promise<any>{
+//		const user: User = await this.service.getUserByLogin(login);
+//		const payloadToSign = {login: login, id: user.id, role: user.userRole}
+//		const access_token = await this.jwtService.signAsync(payloadToSign);
+//		const decoded: JwtPayload = this.jwtService.decode(access_token) as JwtPayload;
+//		return {
+//			access_token: access_token,
+//			expires_at: decoded.exp * 1000, //ms
+//		};
+//	}
 
 	@UseGuards(AuthGuard)
 	@Post('check-username')
 	public async checkUsername(@Body() body: { username: string }): Promise<{ isValid: boolean }> {
 		const { username } = body;
 
+		if (!username) return { isValid: false }
 		const isValidLocal = ValidationFunctions.UsernameValidator(username); // Validación en la lista local
 		if (!isValidLocal) {
 			return { isValid: false }; // Si el nombre de usuario no es válido en la lista local, devolver false
@@ -289,10 +324,11 @@ export class UserController {
 		})
 	}))
 	uploadFile(@Body() body: any, @UploadedFile() file: Express.Multer.File) {
-		console.log(file)
+		if (!file) return false;
 		if (body.fileValidationError) {
 			return { error: body.fileValidationError };
 		  }
+
 		  return { image: file.filename };
 	}
